nomination done

+++ Pages/Students/StudentAssociation.cshtml.cs
@@ -43,13 +43,19 @@ public class StudentAssociation : StudentBasePageModel
     {
         var rollNumber = _context.Students.Single(e => e.EmailId.ToLower() == User.Identity!.Name!.ToLower()).RollNumber;
         var studentAssociation = _context.StudentAssociations.Where(e => e.RollNumber == rollNumber && e.AssociationId == associationId);
+        var isElectionInProgress = _context.Elections.Any(e => e.AssociationId == associationId && e.VotingEnd > DateTime.Now);
+        if (isElectionInProgress)
+        {
+            return RedirectToPage("./ElectionIsInProgress");
+        }
+        
         if (studentAssociation.Any())
         {
             _context.StudentAssociations.RemoveRange(studentAssociation);
         }
         else
         {
-            _context.StudentAssociations.Add(new global::StudentAssociation() { RollNumber = rollNumber, AssociationId = associationId });
+            _context.StudentAssociations.Add(new global::SjcVotersPortal.Data.Models.StudentAssociation() { RollNumber = rollNumber, AssociationId = associationId });
         }
 
         await _context.SaveChangesAsync();
